Discord bot with test commands and a level system
Commands are grabbed using a recursive search of the commands directory. The level system is stored in modules and currently only keeps track of the player's level with no way for the user to see their progress. That is next on my todo list and I have started with it in level.js