View level command works.
insert a pattern.jpg image into the app root directory and it will work. I need to find a copyright free one to include for a preview. I used canvas to make a custom level card for each user. Then I send that as an attachment in a defered reply.